requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
sqlmodel==0.0.21
SQLAlchemy==2.0.30
pydantic==2.8.2
python-docx==1.1.2
httpx==0.27.0
jsonschema==4.23.0
itsdangerous==2.2.0
together>=1.2.0
